1.移除partitionGroup时候首先进行逻辑删除，文件修改为.d 2.定期清理.d文件(.d文件存在超过两天) 3.在broker重启的时候会清理所有的.d文件

+++ joyqueue-server/joyqueue-broker-core/src/main/java/org/joyqueue/broker/config/BrokerStoreConfig.java
@@ -181,7 +181,7 @@ public class BrokerStoreConfig {
     public int getStorePhysicalCleanInterval() {
         return PropertySupplier.getValue(propertySupplier, BrokerStoreConfigKey.STORE_PHYSICAL_CLEAN_INTERVAL, DEFAULT_STORE_PHYSICAL_CLEAN_INTERVAL);
     }
-    public int getStoreDeleteRetainInterval() {
+    public long getStoreDeleteRetainInterval() {
         return PropertySupplier.getValue(propertySupplier, BrokerStoreConfigKey.STORE_DELETE_RETAIN_INTERVAL, DEFAULT_STORE_DELETE_RETAIN_INTERVAL);
     }
